src/requirements.txt
selenium
python-dotenv
pandas
PyPDF2